+++ packages/aws-amplify-react-native/src/Auth/withOAuth.js
@@ -40,6 +40,7 @@ export default (Comp) => {
             this.state = {
                 user: null,
                 error: null,
+                loading: false,
             };
 
             listeners.forEach(listener => Hub.remove('auth', listener));
@@ -50,11 +51,13 @@ export default (Comp) => {
 
         componentDidMount() {
             this._isMounted = true;
-            Auth.currentAuthenticatedUser().then(user => {
-                this.setState({ user })
-            }).catch(error => {
-                logger.debug(error);
-                this.setState({ user: null });
+            this.setState({ loading: true }, () => {
+                Auth.currentAuthenticatedUser().then(user => {
+                    this.setState({ user, loading: false })
+                }).catch(error => {
+                    logger.debug(error);
+                    this.setState({ user: null, loading: false });
+                });
             });
         }
         componentWillUnmount() {
@@ -72,19 +75,19 @@ export default (Comp) => {
                     case 'cognitoHostedUI': {
                         Auth.currentAuthenticatedUser().then(user => {
                             logger.debug('signed in');
-                            this.setState({ user, error: null });
+                            this.setState({ user, error: null, loading: false });
                         });
                         break;
                     }
                     case 'signOut': {
                         logger.debug('signed out');
-                        this.setState({ user: null, error: null });
+                        this.setState({ user: null, error: null, loading: false });
                         break;
                     }
                     case 'signIn_failure':
                     case 'cognitoHostedUI_failure': {
                         logger.debug('not signed in');
-                        this.setState({ user: null, error: decodeURIComponent(payload.data) });
+                        this.setState({ user: null, error: decodeURIComponent(payload.data), loading: false });
                         break;
                     }
                     default:
@@ -108,7 +111,7 @@ export default (Comp) => {
         }
 
         hostedUISignIn(provider) {
-            Auth.federatedSignIn({ provider });
+            this.setState({ loading: true },  () => Auth.federatedSignIn({ provider }));
         }
 
         signOut() {
@@ -116,10 +119,11 @@ export default (Comp) => {
         }
 
         render() {
-            const { user: oAuthUser, error: oAuthError } = this.state;
+            const { user: oAuthUser, error: oAuthError, loading } = this.state;
             const { oauth_config: _, ...otherProps } = this.props;
 
             const oAuthProps = {
+                loading,
                 oAuthUser,
                 oAuthError,
                 hostedUISignIn: this.hostedUISignIn.bind(this, CognitoHostedUIIdentityProvider.Cognito),