+++ packages/api/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.0.35"></a>
+## [1.0.35](https://github.com/aws/aws-amplify/compare/@aws-amplify/api@1.0.35-unstable.2...@aws-amplify/api@1.0.35) (2019-05-14)
+
+
+
+
+**Note:** Version bump only for package @aws-amplify/api
+
+<a name="1.0.35-unstable.2"></a>
+## [1.0.35-unstable.2](https://github.com/aws/aws-amplify/compare/@aws-amplify/api@1.0.35-unstable.1...@aws-amplify/api@1.0.35-unstable.2) (2019-05-14)
+
+
+
+
+**Note:** Version bump only for package @aws-amplify/api
+
+<a name="1.0.35-unstable.1"></a>
+## [1.0.35-unstable.1](https://github.com/aws/aws-amplify/compare/@aws-amplify/api@1.0.35-unstable.0...@aws-amplify/api@1.0.35-unstable.1) (2019-05-13)
+
+
+
+
+**Note:** Version bump only for package @aws-amplify/api
+
 <a name="1.0.35-unstable.0"></a>
 ## [1.0.35-unstable.0](https://github.com/aws/aws-amplify/compare/@aws-amplify/api@1.0.34...@aws-amplify/api@1.0.35-unstable.0) (2019-05-09)